- Update xml file to be current with files I possess - Fix a bug in the disksim module dealing with count sizes - Add test code to make debugging easier in the future - Add parsing for names negotiated with Sameer to be compatible with his version
TODO:  Add and test other interfaces for functions added by Sameer last week.

--miwk

+++ sst/elements/sstdisksim/sstdisksim_event.h
@@ -21,7 +21,7 @@ class sstdisksim_event : public SST::Event {
 public:
   eventtype etype;
   long long pos;
-  long long count;
+  int count;
   unsigned long devno;
   unsigned long time;
